Renamed update_user_signal.py to update_user_assignment.py
